requirements.txt
appdirs==1.4.4
atomicwrites==1.4.0
attrs==20.3.0
CacheControl==0.12.6
cachy==0.3.0
certifi==2020.12.5
chardet==4.0.0
cleo==0.8.1
clikit==0.6.2
colorama==0.4.4
colorclass==2.2.0
crashtest==0.3.1
distlib==0.3.1
docopt==0.6.2
Faker==6.5.0
filelock==3.0.12
html5lib==1.1
idna==3.1
importlib-metadata==3.7.0
iniconfig==1.1.1
keyring==21.8.0
lockfile==0.12.2
more-itertools==8.7.0
msgpack==1.0.2
packaging==20.9
pastel==0.2.1
pendulum==2.1.2
pexpect==4.8.0
pillow==8.1.1
pip-upgrader==1.4.15
pkginfo==1.7.0
pluggy==0.13.1
poetry==1.1.4
poetry-core==1.0.2
ptyprocess==0.7.0
py==1.10.0
pylev==1.3.0
pyparsing==2.4.7
pytest==6.2.2
pytest-rerunfailures==9.1.1
python-dateutil==2.8.1
pytzdata==2020.1
pywin32==300
pywin32-ctypes==0.2.0
radar==0.3
requests==2.25.1
requests-toolbelt==0.9.1
selenium==3.141.0
shellingham==1.4.0
six==1.15.0
terminaltables==3.1.0
text-unidecode==1.3
toml==0.10.2
tomlkit==0.7.0
urllib3==1.26.3
virtualenv==20.4.2
wcwidth==0.2.5
webencodings==0.5.1
zipp==3.4.0
